ㅊㅊ
ㅊㅊ

+++ ONEFIT/src/main/java/org/zerock/b01/dto/All_MemberDTO.java
@@ -21,7 +21,7 @@ public class All_MemberDTO {
 
     private String aPsw;
 
-    private int aPhone;
+    private Long aPhone;
 
     private String memberType;
